requirements.txt
pydantic>=2.0
pydantic-settings
python-dotenv
aiohttp
tenacity
cryptography
watchdog
hvac
redis
aiosqlite
aiofiles
backoff
python-json-logger
prometheus-client
structlog
psutil